Idempotency for Chrome

+++ provisioning/ubuntudev/bootstrap.sh
@@ -58,9 +58,11 @@ function install_vagrant {
 }
 
 function install_chrome {
-	apt-get -y install libxss1
-	wget --output-document=/tmp/vagrant/google-chrome-stable_current_amd64.deb https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-	dpkg -i /tmp/vagrant/google-chrome-stable_current_amd64.deb
+	if [ ! -e /opt/google/chrome/google-chrome ]; then
+		apt-get -y install libxss1
+		wget --output-document=/tmp/vagrant/google-chrome-stable_current_amd64.deb https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
+		dpkg -i /tmp/vagrant/google-chrome-stable_current_amd64.deb
+	fi
 }
 
 initial_setup
